:ok_hand: Added data model.
